Fix depthchart number vs BigDecimal problem.
Ewelina found this bug: when clicking on first buy order on order list, the deptchart
didn't have the summary. It was becouse number and BigNumber where compared by changing
number into BigNumber. When we compare number and BigNumber.toNumber() it works fine (the precision and precision's lost are the same)